+++ packages/app/gen/android/app/build.gradle.kts
@@ -16,10 +16,10 @@ val tauriProperties = Properties().apply {
 
 android {
     compileSdk = 34
-    namespace = "com.ratel.ratelMobile"
+    namespace = "foundation.ratel.app"
     defaultConfig {
         manifestPlaceholders["usesCleartextTraffic"] = "false"
-        applicationId = "com.ratel.ratelMobile"
+        applicationId = "foundation.ratel.app"
         minSdk = 24
         targetSdk = 34
         versionCode = tauriProperties.getProperty("tauri.android.versionCode", "1").toInt()
