+++ packages/app/tauri.conf.json
@@ -2,7 +2,7 @@
   "$schema": "https://schema.tauri.app/config/2",
   "productName": "ratel",
   "version": "0.1.0",
-  "identifier": "com.ratel.ratelMobile",
+  "identifier": "foundation.ratel.app",
   "build": {
     "beforeDevCommand": "cd main-ui && make run",
     "devUrl": "http://localhost:8080",
